Convert custom draftail js to webpack transpile
