+++ fec/fec/static/js/draftail/Anchor.js
@@ -1,0 +1,66 @@
+import React from 'react';
+import { EditorState, Modifier } from 'draft-js';
+import { slugify } from './utils';
+
+// Creates the entities as soon as it is rendered.
+class AnchorSource extends React.Component {
+  componentDidMount() {
+    const { editorState, entityType, onComplete } = this.props;
+
+    const content = editorState.getCurrentContent();
+    const selection = editorState.getSelection();
+
+    // Gets the selected text from the editor
+    var selectionState = editorState.getSelection();
+    var anchorKey = selectionState.getAnchorKey();
+    var currentContentBlock = content.getBlockForKey(anchorKey);
+    var start = selectionState.getStartOffset();
+    var end = selectionState.getEndOffset();
+    var selectedText = currentContentBlock.getText().slice(start, end);
+
+    // Uses the Draft.js API to create a new entity with the right data.
+    const contentWithEntity = content.createEntity(entityType.type, 'IMMUTABLE', {
+        anchor: slugify(selectedText),
+    });
+    const entityKey = contentWithEntity.getLastCreatedEntityKey();
+
+    // Add some text for the entity to be activated on.
+    const text = `${selectedText}`;
+
+    const newContent = Modifier.replaceText(content, selection, text, null, entityKey);
+    const nextState = EditorState.push(editorState, newContent, 'insert-characters');
+
+    onComplete(nextState);
+  }
+
+  render() {
+    return null;
+  }
+}
+
+module.exports.AnchorSource = AnchorSource;
+
+// This adds additional 'term' class to the editor
+// to add custom editor styles inside customize-editor.css
+const Anchor = (props) => {
+  const { entityKey, contentState } = props;
+
+  return React.createElement('span', {
+    style: {
+      fontSize: '2.4rem',
+      margin: '0 0 1em 0',
+      fontFamily: 'gandhi,serif',
+      fontSize: '2rem',
+      fontWeight: '700',
+      lineHeight: '1.2'
+    }
+  }, props.children);
+};
+
+module.exports.Anchor = Anchor;
+
+module.exports = {
+  type: 'ANCHOR',
+  source: AnchorSource,
+  decorator: Anchor
+};
