+++ fec/fec/static/js/draftail/App.js
@@ -1,0 +1,10 @@
+const React = require('react');
+const $ = require('jquery');
+
+const anchor = require('./Anchor');
+const glossary = require('./Glossary');
+const sansserif = require('./SansSerif');
+
+window.draftail.registerPlugin(anchor);
+window.draftail.registerPlugin(glossary);
+window.draftail.registerPlugin(sansserif);
