+++ fec/fec/static/js/draftail/utils.js
@@ -1,0 +1,10 @@
+function slugify(text) {
+  return text.toString().toLowerCase()
+    .replace(/\s+/g, '-')
+    .replace(/[^\w\-]+/g, '')
+    .replace(/\-\-+/g, '-')
+    .replace(/^-+/, '')
+    .replace(/-+$/, '');
+}
+
+module.exports.slugify = slugify;
