+++ fec/fec/wagtail_hooks.py
@@ -1,8 +1,10 @@
+import json
 import wagtail.admin.rich_text.editors.draftail.features as draftail_features
 from wagtail.admin.rich_text.converters.html_to_contentstate import InlineStyleElementHandler, BlockElementHandler
 from wagtail.admin.rich_text.converters.editor_html import WhitelistRule
 from wagtail.core import hooks
 from wagtail.core.whitelist import attribute_rule, check_url, allow_without_attributes
+from django.conf import settings
 from django.utils.html import format_html
 from fec.draftail import glossary, sansserif, anchor
 
@@ -116,15 +118,26 @@ def register_sansserif_feature(features):
         'to_database_format': {'entity_decorators': {type_: sansserif.sansserif_entity_decorator}},
     })
 
+
+def draftail_js():
+    """Looks up the hashed asset path in rev-draftail-manifest-js.json
+    If the path doesn't exist there, then just return the path to the static file
+    without a hash"""
+    key = '/static/js/draftail.js'
+    assets = json.load(open(settings.DIST_DIR + '/fec/static/js/rev-draftail-manifest-js.json'))
+    return assets[key] if key in assets else key
+
 # Inserts custom editor js
 @hooks.register('insert_editor_js')
 def editor_js():
-    return format_html('''
-        <script src="/static/wagtailadmin/js/draftail.js"></script>
-        <script src="/static/js/admin/glossary.js"></script>
-        <script src="/static/js/admin/sansserif.js"></script>
-        <script src="/static/js/admin/anchor.js"></script>
-    ''')
+    transpiled = draftail_js()
+    scripts = [
+        '<script src="/static/wagtailadmin/js/draftail.js"></script>',
+        '<script src="{}"></script>'.format(transpiled)
+    ]
+
+    html = '\n'.join(str(s) for s in scripts)
+    return format_html(html)
 
 @hooks.register('insert_editor_css')
 def editor_css():
