+++ package.json
@@ -23,13 +23,15 @@
     "babel-eslint": "^7.1.1",
     "babel-preset-latest": "^6.24.0",
     "babelify": "^7.3.0",
-    "chroma-js": "1.0.1",
     "bleach": "0.3.0",
+    "chroma-js": "1.0.1",
     "colorbrewer": "0.0.2",
     "component-sticky": "1.0.0",
     "d3": "3.5.5",
     "datatables.net": "1.10.10",
     "datatables.net-responsive": "2.0.1",
+    "draft-js": "^0.10.5",
+    "draftail": "^0.17.2",
     "es6-weak-map": "2.0.1",
     "eventemitter2": "0.4.14",
     "fullcalendar": "3.3.1",
